+++ packages/common/src/signal.ts
@@ -1,8 +1,8 @@
-import { KubeDocument } from '@traxitt/kubeclient'
+import { KubeDocument } from '@c6o/kubeclient'
 
 // annotation added to document to trigger application update
-export const DOCUMENT_SIGNAL = 'system.traxitt.com/update-signal'
-export const DOCUMENT_SIGNAL_JSON_PATCH = 'system.traxitt.com~1update-signal'
+export const DOCUMENT_SIGNAL = 'system.codezero.io/update-signal'
+export const DOCUMENT_SIGNAL_JSON_PATCH = 'system.codezero.io~1update-signal'
 
 export const isDocumentSignalled = (document: KubeDocument) => document.metadata?.annotations?.[DOCUMENT_SIGNAL]  === 'true'
 export const signalDocument = (document: KubeDocument) => {