+++ packages/grafana/src/ui/install.ts
@@ -15,14 +15,14 @@ export class GrafanaCredentials extends LitElement implements StoreFlowStep {
 
     render() {
         return html`
-            <traxitt-form-layout>
-                <traxitt-combo-box @selected-item-changed=${this.storageSelected} label="Grafana Storage" value=${this.serviceSpec.storage} required allow-custom-value .items=${this.values}></traxitt-combo-box>
-            </traxitt-form-layout>
-            <traxitt-form-layout>
-                <traxitt-text-field @input=${this.usernameChanged} label="Administrator username" value=${this.serviceSpec.adminUsername} autoselect required></traxitt-text-field>
+            <c6o-form-layout>
+                <c6o-combo-box @selected-item-changed=${this.storageSelected} label="Grafana Storage" value=${this.serviceSpec.storage} required allow-custom-value .items=${this.values}></c6o-combo-box>
+            </c6o-form-layout>
+            <c6o-form-layout>
+                <c6o-text-field @input=${this.usernameChanged} label="Administrator username" value=${this.serviceSpec.adminUsername} autoselect required></c6o-text-field>
                 <br />
-                <traxitt-text-field @input=${this.passwordChanged} label="Administrator password" value=${this.serviceSpec.adminPassword} autoselect required></traxitt-text-field>
-            </traxitt-form-layout>
+                <c6o-text-field @input=${this.passwordChanged} label="Administrator password" value=${this.serviceSpec.adminPassword} autoselect required></c6o-text-field>
+            </c6o-form-layout>
         `
     }