+++ packages/istio/src/index.ts
@@ -1,6 +1,5 @@
-import { mix } from "mixwith"
-import { ProvisionerBase } from "@provisioner/common"
-
+import { mix } from 'mixwith'
+import { ProvisionerBase } from '@provisioner/common'
 import {
     removeApplyMixin,
     createInquireMixin,
@@ -30,3 +29,5 @@ export class Provisioner extends mix(ProvisionerBase).with(
     httpsRedirectApiMixin,
     virtualServiceApiMixin) {
 }
+
+export * from './constants'