+++ packages/mattermost/src/index.ts
@@ -3,10 +3,17 @@ import { ProvisionerBase } from "@provisioner/common"
 
 import {
     createApplyMixin,
-    createInquireMixin
+    createInquireMixin,
+    createValidateMixin,
+    removeApplyMixin,
+    removeInquireMixin
 } from './mixins'
 
 export type baseProvisionerType = new (...a) => Provisioner & ProvisionerBase
 
-export class Provisioner extends mix(ProvisionerBase).with(createApplyMixin, createInquireMixin) {
+export interface Provisioner extends ProvisionerBase {
+}
+
+export class Provisioner extends mix(ProvisionerBase).with(createApplyMixin, createInquireMixin, createValidateMixin, removeApplyMixin, removeInquireMixin) {
+    get isPreview() { return this.edition === 'preview' }
 }