+++ packages/logging-elk/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.5](https://github.com/c6o/provisioners/compare/v0.0.4...v0.0.5) (2020-08-13)
+
+
+### ♻️ Chores
+
+* Updated gitHead ([570855f](https://github.com/c6o/provisioners/commit/570855fb1f45f0e051dedccc2acef7b83375ebac))
+
+
+
+
+
+## [0.0.4](https://github.com/c6o/provisioners/compare/v0.0.3...v0.0.4) (2020-08-10)
+
+
+### ♻️ Chores
+
+* Updated gitHead ([2045140](https://github.com/c6o/provisioners/commit/2045140b6ae8bc2e4504ff7756b7a8776c087609))
+
+
+
+
+
+## [0.0.3](https://github.com/c6o/provisioners/compare/v0.0.2...v0.0.3) (2020-08-08)
+
+
+### 🐛 Bug Fixes
+
+* **NavStation:** Improvements to Nav Station UI ([#14](https://github.com/c6o/provisioners/issues/14)) ([fccaf70](https://github.com/c6o/provisioners/commit/fccaf7057be6de5235267fe0bbf6dc5be29e583f))
+
+
+### ♻️ Chores
+
+* Bumped kubeclient to 0.0.5 ([ba51a57](https://github.com/c6o/provisioners/commit/ba51a574b2a123bbe012be0086ec2ecbedcf487c))
+* Licensing ([#17](https://github.com/c6o/provisioners/issues/17)) ([8b9cc24](https://github.com/c6o/provisioners/commit/8b9cc24ff42ff875b4234a74dfcfcfedb2acef27))
+* Licensing and repo links ([#16](https://github.com/c6o/provisioners/issues/16)) ([b8b03bb](https://github.com/c6o/provisioners/commit/b8b03bbe7f30904b83cc599e61d378beb009eb38))
+* Marked packages public ([39a328e](https://github.com/c6o/provisioners/commit/39a328e0225b2b773e173960f54f98052a698368))
+* Move all our packages to the same version of lit-element ([#15](https://github.com/c6o/provisioners/issues/15)) ([013dbd6](https://github.com/c6o/provisioners/commit/013dbd6377a1f52f5a3a71885e7935e0c4984a21))
+* UI cleanup for logging-elk provisioner ([#13](https://github.com/c6o/provisioners/issues/13)) ([12e6b27](https://github.com/c6o/provisioners/commit/12e6b27070609762bcbe110a7ec3cb145d97e3f1))
+
+
+
+
+
 ## [0.0.2](https://github.com/traxitt/traxitt/compare/v0.0.1...v0.0.2) (2020-07-17)
 
 **Note:** Version bump only for package @provisioner/logging-elk